Introduced resource uniqueness for GPU types
+ This API change forbids, at compile time, the usage of GPU resources
in other GPUs other than the owner of said resouce.
+ For example, if GPU1 owns a buffer to its own memory, GPU2 won't be
able to use it except for the designated transfer operations (yet to be
implemented). Since the GPUs are only known at runtime, with the
creation of a unique type (unnamed and opaque type) as a tag for the
resouces of the same GPU, we can effectively limit who and how those
resources are used.

+ Only `simple-compute` example has been ported. Updating the rest
should be fairly simple.

+ https://github.com/Rimpampa/unique-type/issues/3. Here you got an
example!